[refactor] Segment 이름 수정
