+++ src/features/region/index.ts
@@ -1,1 +1,2 @@
 export type { Position, Region } from './model/types';
+export * from './model/useRegionStore';
